move SSLContextProvider and implementing classes to new package
