+++ src/main/java/org/cryptomator/launcher/Cryptomator.java
@@ -11,7 +11,7 @@ import org.apache.commons.lang3.SystemUtils;
 import org.cryptomator.common.Environment;
 import org.cryptomator.common.ShutdownHook;
 import org.cryptomator.common.SubstitutingProperties;
-import org.cryptomator.common.sslcontext.SSLContextProvider;
+import org.cryptomator.common.integrations.sslcontext.SSLContextProvider;
 import org.cryptomator.ipc.IpcCommunicator;
 import org.cryptomator.logging.DebugMode;
 import org.cryptomator.ui.fxapp.FxApplicationComponent;
